Add scoresheet naming preference, configure scoresheet download based upon preference, bug fix QR code function, begin address of output function errors.

+++ admin/entries.admin.php
@@ -73,23 +73,24 @@ do {
 	$required_info = "";
 	$entry_judging_num = "";
 	$entry_judging_num_display = "";
-	
+
 	$entry_number = sprintf("%04s",$row_log['id']);
 	$judging_number = sprintf("%06s",$row_log['brewJudgingNumber']);
-	
+
 	// Check whether scoresheet file exists, and, if so, provide link.
 	$scoresheet_file_name_entry = sprintf("%06s",$entry_number).".pdf";
 	$scoresheet_file_name_judging = $judging_number.".pdf";
 	$scoresheetfile_entry = USER_DOCS.$scoresheet_file_name_entry;
 	$scoresheetfile_judging = USER_DOCS.$scoresheet_file_name_judging;
-	
+
 	$scoresheet = FALSE;
 	if ((file_exists($scoresheetfile_entry)) || (file_exists($scoresheetfile_judging))) $scoresheet = TRUE;
-	
-	if (file_exists($scoresheetfile_entry)) $scoresheet_file_name = $scoresheet_file_name_entry;
-	elseif (file_exists($scoresheetfile_judging)) $scoresheet_file_name = $scoresheet_file_name_judging;
-	else $scoresheet_file_name = "";
-	
+
+	$scoresheet_file_name_1 = "";
+	$scoresheet_file_name_2 = "";
+	if (file_exists($scoresheetfile_entry)) $scoresheet_file_name_1 = $scoresheet_file_name_entry;
+	if (file_exists($scoresheetfile_judging)) $scoresheet_file_name_2 = $scoresheet_file_name_judging;
+
 	$filename_judging = USER_DOCS.$judging_number.".pdf";
 	if (file_exists($filename_judging)) $scoresheet = TRUE;
 
@@ -197,44 +198,76 @@ do {
 		$entry_actions .= "<a href=\"mailto:".$brewer_info[6]."\" data-toggle=\"tooltip\" data-placement=\"top\" title=\"Email the entry&rsquo;s owner, ".$brewer_info[0]." ".$brewer_info[1].", at ".$brewer_info[6]."\"><span class=\"fa fa-lg fa-envelope\"></span></a> ";
 	}
 
+	$scoresheet_link_1 = "";
+	$scoresheet_link_2 = "";
+
 	if ($scoresheet) {
-		
-		// The pseudo-random number and the corresponding name of the temporary file are defined each time 
-		// this brewer_entries.sec.php script is accessed (or refreshed), but the temporary file is created
-		// only when the entrant clicks on the gavel icon to access the scoresheet. 
-		$random_num_str = random_generator(8,2);
-		$random_file_name = $random_num_str.".pdf";
-		$scoresheet_random_file_relative = "user_temp/".$random_file_name;
-		$scoresheet_random_file = USER_TEMP.$random_file_name;
-		$scoresheet_random_file_html = $base_url.$scoresheet_random_file_relative;
 
-		if (!empty($scoresheet_file_name)) {
-			$scoresheet_link = "";
-			$scoresheet_link .= "<a href=\"".$base_url."output/scoresheets.output.php?";
+		if (!empty($scoresheet_file_name_1))  {
 
-			// Obfuscate the *ACTUAL* file names. 
+			// The pseudo-random number and the corresponding name of the temporary file are defined each time
+			// this brewer_entries.sec.php script is accessed (or refreshed), but the temporary file is created
+			// only when the entrant clicks on the gavel icon to access the scoresheet.
+			$random_num_str_1 = random_generator(8,2);
+			$random_file_name_1 = $random_num_str_1.".pdf";
+			$scoresheet_random_file_relative_1 = "user_temp/".$random_file_name_1;
+			$scoresheet_random_file_1 = USER_TEMP.$random_file_name_1;
+			$scoresheet_random_file_html_1 = $base_url.$scoresheet_random_file_relative_1;
+
+
+			$scoresheet_link_1 .= "<a href=\"".$base_url."output/scoresheets.output.php?";
+
+			// Obfuscate the *ACTUAL* file names.
 			// Prevents casual users from right clicking on scoresheet download link and changing
-			// the entry or judging number pdf name passed via the URL to force downloads of files 
+			// the entry or judging number pdf name passed via the URL to force downloads of files
 			// they shouldn't have access to. Can I get a harumph?!
-			$scoresheet_link .= "scoresheetfilename=".encryptString($scoresheet_file_name);
-			$scoresheet_link .= "&amp;randomfilename=".encryptString($random_file_name)."&amp;download=true";
-			$scoresheet_link .= sprintf("\" data-toggle=\"tooltip\" title=\"%s '".$row_log['brewName']."'.\">",$brewer_entries_text_006);
-			$scoresheet_link .= "<span class=\"fa fa-lg fa-gavel\"></a>&nbsp;&nbsp;";
+			$scoresheet_link_1 .= "scoresheetfilename=".encryptString($scoresheet_file_name_1);
+			$scoresheet_link_1 .= "&amp;randomfilename=".encryptString($random_file_name_1)."&amp;download=true";
+			$scoresheet_link_1 .= sprintf("\" data-toggle=\"tooltip\" title=\"%s '".$row_log['brewName']."'' (by Entry Number).\">",$brewer_entries_text_006);
+			$scoresheet_link_1 .= "<span class=\"fa fa-lg fa-gavel\"></a>&nbsp;&nbsp;";
+		}
+
+		if (!empty($scoresheet_file_name_2))  {
+
+			// The pseudo-random number and the corresponding name of the temporary file are defined each time
+			// this brewer_entries.sec.php script is accessed (or refreshed), but the temporary file is created
+			// only when the entrant clicks on the gavel icon to access the scoresheet.
+
+			$random_num_str_2 = random_generator(8,2);
+			$random_file_name_2 = $random_num_str_2.".pdf";
+			$scoresheet_random_file_relative_2 = "user_temp/".$random_file_name_2;
+			$scoresheet_random_file_2 = USER_TEMP.$random_file_name_2;
+			$scoresheet_random_file_html_2 = $base_url.$scoresheet_random_file_relative_2;
+
+			$scoresheet_link_2 .= "<a href=\"".$base_url."output/scoresheets.output.php?";
+
+			// Obfuscate the *ACTUAL* file names.
+			// Prevents casual users from right clicking on scoresheet download link and changing
+			// the entry or judging number pdf name passed via the URL to force downloads of files
+			// they shouldn't have access to. Can I get a harumph?!
+			$scoresheet_link_2 .= "scoresheetfilename=".encryptString($scoresheet_file_name_2);
+			$scoresheet_link_2 .= "&amp;randomfilename=".encryptString($random_file_name_2)."&amp;download=true";
+			$scoresheet_link_2 .= sprintf("\" data-toggle=\"tooltip\" title=\"%s '".$row_log['brewName']."' (by Judging Number).\">",$brewer_entries_text_006);
+			$scoresheet_link_2 .= "<span class=\"fa fa-lg fa-gavel\"></a>&nbsp;&nbsp;";
 		}
 
 		// Clean up temporary scoresheets created for other brewers, when they are at least 1 minute old (just to avoid problems when two entrants try accessing their scoresheets at practically the same time, and clean up previously created scoresheets for the same brewer, regardless of how old they are.
 		$tempfiles = array_diff(scandir(USER_TEMP), array('..', '.'));
-		foreach ($tempfiles as $file) {
-			if ((filectime(USER_TEMP.$file) < time() - 1*60) || ((strpos($file, $scoresheet_file_name_judging) !== FALSE))) {
-				unlink(USER_TEMP.$file);
-			}
+		
+		if (is_array($tempfiles)) {
+			foreach ($tempfiles as $file) {
+				if ((filectime(USER_TEMP.$file) < time() - 1*60) || ((strpos($file, $scoresheet_file_name_judging) !== FALSE))) {
+					unlink(USER_TEMP.$file);
+				}
 
-			if ((filectime(USER_TEMP.$file) < time() - 1*60) || ((strpos($file, $scoresheet_file_name_entry) !== FALSE))) {
-				unlink(USER_TEMP.$file);
+				if ((filectime(USER_TEMP.$file) < time() - 1*60) || ((strpos($file, $scoresheet_file_name_entry) !== FALSE))) {
+					unlink(USER_TEMP.$file);
+				}
 			}
 		}
-		
-		$entry_actions .= $scoresheet_link;
+
+		if ((($dbTable == "default") && ($_SESSION['prefsDisplaySpecial'] == "E")) || ($dbTable != "default")) $entry_actions .= $scoresheet_link_1;
+		if ((($dbTable == "default") && ($_SESSION['prefsDisplaySpecial'] == "J")) || ($dbTable != "default")) $entry_actions .= $scoresheet_link_2;
 	}
 
 	$tbody_rows .= "<tr class=\"".$entry_unconfirmed_row."\">";
